# gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/resources/openapi/management-openapi-v2.yaml
openapi: 3.0.3
info:
    title: Gravitee.io APIM - Management API
    description: |-
        This is the OpenAPI specification for our new version of APIM Management API.
    contact:
        email: team-apim@graviteesource.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    version: 2.0.0
security:
    - BasicAuth: []
    - CookieAuth: []
servers:
    - url: /management/v2
      description: Gravitee.io APIM - Management API -v2
...
paths:
...
    /environments/{envId}/apis/{apiId}/pages:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/apiIdParam"
        get:
            tags:
                - API Documentation
            summary: Get API documentation pages
            description: |-
                Get API documentation pages. The result contains a list of pages and folders of the API.
                
                User must have the API_DOCUMENTATION[READ] permission.
            operationId: getApiPages
            responses:
                "200":
                    $ref: "#/components/responses/ApiDocumentationPagesResponse"
                default:
                    $ref: "#/components/responses/Error"
        post:
            tags:
                - API Documentation
            summary: Create API documentation page
            description: |-
                Create API documentation. The result will be the created page or folder.
                Only MARKDOWN and FOLDER types are supported.

                User must have the API_DOCUMENTATION[CREATE] permission.
            operationId: createDocumentation
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateDocumentation"
                required: true
            responses:
                "201":
                    description: Created API Documentation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page"
                default:
                    $ref: "#/components/responses/Error"


    # Plugins - Endpoints
...
components:
    schemas:
...
        Visibility:
            type: string
            description: The visibility of the resource regarding the portal.
            example: PUBLIC
            enum:
                - PUBLIC
                - PRIVATE
...
        Error:
            type: object
            properties:
                httpStatus:
                    type: integer
                    format: int32
                    description: The error code
                    example: 400
                message:
                    type: string
                    description: The error message
                    example: Bad request
                technicalCode:
                    type: string
                    description: A technical code to identify the error
                    example: invalid.import.definition
                parameters:
                    type: object
                    description: A map of parameters to be used in the error message
                    additionalProperties:
                        type: string
                details:
                    type: array
                    description: A list of details about the error
                    items:
                        type: object
                        properties:
                            message:
                                type: string
                                description: The error message
                                example: Bad request
                            location:
                                type: string
                                description: The json path of the field in error.
                                example: updateApi.properties[0].key
                            invalidValue:
                                type: object
                                description: The invalid value.
...
        AccessControl:
            type: object
            properties:
                referenceId:
                    type: string
                    description: The id of the resource used to check the access control
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                referenceType:
                    type: string
                    description: The type of the resource used to check the access control
                    example:
                        - GROUP
                        - ROLE
...
        Page:
            description: Documentation page. Can be of multiple formats.
            type: object
            properties:
                id:
                    type: string
                    description: Page's uuid.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                crossId:
                    type: string
                    description: Page's cross uuid.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                name:
                    type: string
                    description: Page's name.
                    example: My Page
                type:
                    $ref: "#/components/schemas/PageType"
                content:
                    type: string
                    description: Page's content.
                    example: My Page content
                order:
                    type: integer
                    description: Page's order.
                    example: 1
                lastContributor:
                    type: string
                    description: Page's last contributor. Id of a user.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                published:
                    type: boolean
                    description: Page's published status.
                    example: true
                visibility:
                    $ref: "#/components/schemas/Visibility"
                updatedAt:
                    type: string
                    format: date-time
                    description: Page's last update date.
                    example: 2018-01-01T00:00:00.000Z
                contentType:
                    type: string
                    description: Page's content type.
                    example: application/json
                source:
                    $ref: "#/components/schemas/PageSource"
                configuration:
                    type: object
                    description: Page's configuration.
                    additionalProperties:
                        type: string
                homepage:
                    type: boolean
                    description: Page's homepage status.
                    example: true
                parentId:
                    type: string
                    description: Page's parent id.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                parentPath:
                    type: string
                    description: Page's parent path.
                    example: /parent
                excludedGroups:
                    type: array
                    description: List of group ids that are considered when looking for acces control.
                    items:
                        type: string
                        example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                excludedAccessControls:
                    type: boolean
                    description: Flag to restrict access to user matching the restrictions.
                accessControls:
                    type: array
                    description: List of access controls.
                    items:
                        $ref: "#/components/schemas/AccessControl"
                attachedMedia:
                    type: array
                    description: List of attached media.
                    items:
                        $ref: "#/components/schemas/PageMedia"
                metadata:
                    type: object
                    description: Page's metadata.
                    additionalProperties:
                        type: string
                translations:
                    type: array
                    description: List of page translations.
                    items:
                        $ref: "#/components/schemas/Page"
                generalConditions:
                    type: boolean
                    description: Is the page used as General Conditions for a plan.
                contentRevision:
                    $ref: "#/components/schemas/Revision"
        PageMedia:
            type: object
            properties:
                name:
                    type: string
                    description: Media's name.
                    example: My Media
                hash:
                    type: string
                    description: Media's hash.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                attachedAt:
                    type: string
                    format: date-time
                    description: Media's attachment date.
                    example: 2018-01-01T00:00:00.000Z
        PageSource:
            type: object
            properties:
                type:
                    type: string
                    description: The type of the page source (=fetcher type).
                    example:
                        - http-fetcher
                        - git-fetcher
                        - gitlab-fetcher
                        - github-fetcher
                        - bitbucket-fetcher
                configuration:
                    type: object
                    description: Page source's configuration.
        PageType:
            type: string
            description: The type of the page.
            example: MARKDOWN
            enum:
                - ASCIIDOC
                - ASYNCAPI
                - MARKDOWN
                - MARKDOWN_TEMPLATE
                - SWAGGER
                - FOLDER
                - LINK
                - ROOT
                - SYSTEM_FOLDER
                - TRANSLATION
        CreateDocumentation:
            type: object
            oneOf:
                - $ref: "#/components/schemas/CreateDocumentationFolder"
                - $ref: "#/components/schemas/CreateDocumentationMarkdown"
            required:
                - name
                - type
            discriminator:
                propertyName: type
                mapping:
                    FOLDER: CreateDocumentationFolder
                    MARKDOWN: CreateDocumentationMarkdown
        BaseCreateDocumentation:
            type: object
            properties:
                name:
                    type: string
                    description: Page's name.
                    example: My Page
                type:
                    type: string
                    description: The type of the page. Only MARKDOWN and FOLDER types can be created.
                    example: MARKDOWN
                    enum:
                        - MARKDOWN
                        - FOLDER
                order:
                    type: integer
                    description: Page's order.
                    example: 1
                visibility:
                    $ref: "#/components/schemas/Visibility"
                parentId:
                    type: string
                    description: Page's parent id.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
            required:
                - name
                - type
            discriminator:
                propertyName: type
                mapping:
                    FOLDER: CreateDocumentationFolder
                    MARKDOWN: CreateDocumentationMarkdown
        CreateDocumentationFolder:
            title: "CreateDocumentationFolder"
            description: Create documentation folder.
            type: object
            allOf:
                - $ref: "#/components/schemas/BaseCreateDocumentation"
        CreateDocumentationMarkdown:
            title: "CreateDocumentationMarkdown"
            description: Create documentation markdown page.
            type: object
            allOf:
                - $ref: "#/components/schemas/BaseCreateDocumentation"
                - properties:
                    content:
                        type: string
                        description: Page's content.
                        example: My Page content
                    homepage:
                        type: boolean
                        description: Page's homepage status.
                        example: true
                    generalConditions:
                        type: boolean
                        description: Is the page used as General Conditions for a plan.
        Revision:
            type: object
            properties:
                id:
                    type: string
                    description: Id of the page used to fill the content attributes.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                revision:
                    type: integer
                    description: Revision number.
                    example: 1

        # Metadata Page
...
    parameters:
...
        envIdParam:
            name: envId
            in: path
            required: true
            description: Id or Hrid (Human readable Id) of an environment.
            schema:
                type: string
                default: DEFAULT
        apiIdParam:
            name: apiId
            in: path
            required: true
            description: Id of an API.
            schema:
                type: string
...
    responses:
...
        Error:
            description: Generic error response
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
...
        ApiDocumentationPagesResponse:
            description: API documentation pages
            content:
                application/json:
                    schema:
                        title: "ApiDocumentationPagesResponse"
                        properties:
                            pages:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Page"
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
        CookieAuth:
            type: apiKey
            in: cookie
            name: Auth-Graviteeio-APIM